pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paystack-client"
version = "1.0.0"
description = "A Python wrapper for the Paystack API"
authors = [
    {name = "Joseph Ezekiel", email = "theolujay@gmail.com"},
]
maintainers = [
    {name = "Joseph Ezekiel", email = "theolujay@gmail.com"},
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
requires-python = ">=3.9"  # Changed from >=3.8
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Django",
]
keywords = ["paystack", "payment", "api", "django", "python", "fintech", "nigeria"]

dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "responses>=0.25.8",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]
test = [
    "pytest>=8.4.1", 
    "responses>=0.25.8",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/theolujay/paystack-client"
Repository = "https://github.com/theolujay/paystack-client"
Documentation = "https://paystack-client.readthedocs.io"
"Bug Tracker" = "https://github.com/theolujay/paystack-client/issues"

[tool.hatch.build.targets.wheel]
packages = ["paystack_client"]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = ["--tb=short", "--cov=paystack_client", "--cov-report=term-missing"]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py39'] 

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "N", "W"]

[dependency-groups]
dev = [
    "build>=1.3.0",
    "twine>=6.1.0",
]
